cleaner code

+++ src/main/java/tech/sobhan/golestan/services/CourseService.java
@@ -1,10 +1,10 @@
 package tech.sobhan.golestan.services;
 
 import org.springframework.stereotype.Service;
-import tech.sobhan.golestan.auth.User;
 import tech.sobhan.golestan.business.exceptions.CourseNotFoundException;
 import tech.sobhan.golestan.models.Course;
 import tech.sobhan.golestan.repositories.CourseRepository;
+import tech.sobhan.golestan.security.ErrorChecker;
 
 import java.util.List;
 
@@ -14,31 +14,35 @@ import static tech.sobhan.golestan.utils.Util.deleteLog;
 @Service
 public class CourseService {
     private final CourseRepository courseRepository;
+    private final ErrorChecker errorChecker;
 
-    public CourseService(CourseRepository courseRepository) {
+    public CourseService(CourseRepository courseRepository,
+                         ErrorChecker errorChecker) {
         this.courseRepository = courseRepository;
+        this.errorChecker = errorChecker;
     }
 
-    public List<Course> list() {
-        List<Course> allCourses = courseRepository.findAll();
-//        if(allCourses.isEmpty()){
-//            throw new CourseNotFoundException();
-//        }
-        return allCourses;
+    public String list(String username, String password) {
+        errorChecker.checkIsUser(username, password);
+        return list().toString();
     }
 
-    public Course create(int units, String title) {
+    private List<Course> list() {
+        return courseRepository.findAll();
+    }
+
+    public String create(int units, String title, String username, String password) {
+        errorChecker.checkIsAdmin(username, password);
         Course course = Course.builder().units(units).title(title).build();
-        return create(course);
+        return create(course).toString();
     }
 
     public Course create(Course course) {
-        if (courseExists(list(), course)) return null;
         createLog(Course.class, course.getId());
         return courseRepository.save(course);
     }
 
-    private boolean courseExists(List<Course> allCourses, Course course) {
+    private boolean courseExists(List<Course> allCourses, Course course) {//todo find usage
         for (Course c : allCourses) {
             if(course.equals(c)){
                 System.out.println("ERROR403 duplicate Courses");
@@ -48,23 +52,36 @@ public class CourseService {
         return false;
     }
 
-    public Course read(Long id) {
+    public String read(Long id, String username, String password) {
+        errorChecker.checkIsUser(username, password);
+        return read(id).toString();
+    }
+
+    private Course read(Long id) {
         return courseRepository.findById(id).orElseThrow(CourseNotFoundException::new);
     }
 
-    public void update(Course newCourse, Long id) {
+    public String update(Course newCourse, Long id, String username, String password) {
+        errorChecker.checkIsAdmin(username, password);
+        update(newCourse, id);
+        return "OK";
+    }
+
+    private void update(Course newCourse, Long id) {
         courseRepository.findById(id).map(user -> {
             user = newCourse.clone();
-            return courseRepository.save(user);//todo sus for saveUser instead of user
+            return courseRepository.save(user);
         }).orElseGet(() -> {
             newCourse.setId(id);
             return courseRepository.save(newCourse.clone());
         });
     }
 
-    public void delete(Long id) {
+    public String delete(Long id, String username, String password) {
+        errorChecker.checkIsAdmin(username, password);
         courseRepository.delete(read(id));
         deleteLog(Course.class, id);
+        return "OK";
     }
