+++ src/main/java/tech/sobhan/golestan/services/InstructorService.java
@@ -2,19 +2,14 @@ package tech.sobhan.golestan.services;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import tech.sobhan.golestan.business.exceptions.CourseNotFoundException;
-import tech.sobhan.golestan.business.exceptions.CourseSectionRegistrationNotEmpty;
-import tech.sobhan.golestan.business.exceptions.InstructorNotFoundException;
-import tech.sobhan.golestan.business.exceptions.StudentNotFoundException;
+import tech.sobhan.golestan.business.exceptions.*;
 import tech.sobhan.golestan.enums.Rank;
 import tech.sobhan.golestan.models.CourseSection;
 import tech.sobhan.golestan.models.CourseSectionRegistration;
 import tech.sobhan.golestan.models.users.Instructor;
 import tech.sobhan.golestan.models.users.Student;
-import tech.sobhan.golestan.repositories.CourseSectionRegistrationRepository;
-import tech.sobhan.golestan.repositories.CourseSectionRepository;
-import tech.sobhan.golestan.repositories.InstructorRepository;
-import tech.sobhan.golestan.repositories.StudentRepository;
+import tech.sobhan.golestan.repositories.*;
+import tech.sobhan.golestan.security.ErrorChecker;
 
 import java.util.List;
 import java.util.Map;
@@ -24,27 +19,38 @@ import static tech.sobhan.golestan.utils.Util.createLog;
 @Slf4j
 @Service
 public class InstructorService {
+
+    private final UserRepository userRepository;
+    private final ErrorChecker errorChecker;
+
     private final InstructorRepository instructorRepository;
     private final StudentRepository studentRepository;
     private final CourseSectionRepository courseSectionRepository;
     private final CourseSectionRegistrationRepository courseSectionRegistrationRepository;
 
-    public InstructorService(InstructorRepository instructorRepository, StudentRepository studentRepository,
-                                CourseSectionRepository courseSectionRepository,
-                                CourseSectionRegistrationRepository courseSectionRegistrationRepository) {
+    public InstructorService(UserRepository userRepository, ErrorChecker errorChecker, InstructorRepository instructorRepository,
+                             StudentRepository studentRepository, CourseSectionRepository courseSectionRepository,
+                             CourseSectionRegistrationRepository courseSectionRegistrationRepository) {
+        this.userRepository = userRepository;
+        this.errorChecker = errorChecker;
         this.instructorRepository = instructorRepository;
         this.studentRepository = studentRepository;
         this.courseSectionRepository = courseSectionRepository;
         this.courseSectionRegistrationRepository = courseSectionRegistrationRepository;
     }
 
-    public List<Instructor> list() {
+    public String list(String username, String password) {
+        errorChecker.checkIsUser(username, password);
         //        if(allInstructors.isEmpty()){
 //            throw new InstructorNotFoundException();
 //        }
+        return list().toString();
+    }
+
+    private List<Instructor> list() {
         return instructorRepository.findAll();
     }
-    
+
     public Instructor create(Instructor instructor){
         if (instructorExists(list(), instructor)) return null;
         createLog(Instructor.class, instructor.getId());
@@ -61,11 +67,17 @@ public class InstructorService {
         return false;
     }
 
-    public Instructor read(Long id) {//todo create service package
+    public String read(Long id, String username, String password) {//todo create service package
+        errorChecker.checkIsUser(username, password);
+        return read(id).toString();
+    }
+
+    private Instructor read(Long id) {
         return instructorRepository.findById(id).orElseThrow(InstructorNotFoundException::new);
     }
 
-    public void update(Rank rank, Long id) {
+    public String update(Rank rank, Long id, String username, String password) {
+        errorChecker.checkIsAdmin(username, password);
         instructorRepository.findById(id).map(instructor -> {
             instructor.setRank(rank);
             return instructorRepository.save(instructor);//todo sus for saveinstructor instead of instructor
@@ -73,29 +85,44 @@ public class InstructorService {
             Instructor newInstructor = Instructor.builder().rank(rank).id(id).build();
             return instructorRepository.save(newInstructor);
         });
+        return "OK";
     }
 
-    public void delete(Long id) {//todo check later
+    public String delete(Long id, String username, String password) {//todo check later
+        errorChecker.checkIsAdmin(username, password);
         Instructor instructor = read(id);
         List<CourseSection> courseSectionsOfInstructor = courseSectionRepository.findByInstructor(id);//todo delete course section
         courseSectionRepository.deleteAll(courseSectionRepository.findByInstructor(id));
         instructorRepository.delete(instructor);
-        log.info("Instructor with id of " + id + "deleted successfully.");        //todo add log in all deletes and signups
+        return "OK";
+//        log.info("Instructor with id of " + id + "deleted successfully.");        //todo add log in all deletes and signups
     }
 
-    public void giveMark(Map<String, String> json) {
-        Student foundStudent = studentRepository.
-                findById(Long.valueOf(json.get("student_id"))).orElseThrow(StudentNotFoundException::new);
-        CourseSection foundCourseSection = courseSectionRepository
+    public String giveMark(Map<String, String> json, String username, String password) {
+        CourseSection courseSection = courseSectionRepository
                 .findById(Long.valueOf(json.get("course_section_id"))).orElseThrow(CourseNotFoundException::new);
+        String foundError = foundError(username, password, courseSection);
+        if (foundError != null) return foundError;
+        Student student = studentRepository
+                .findById(Long.valueOf(json.get("student_id"))).orElseThrow(StudentNotFoundException::new);
+
         double score = Double.parseDouble(json.get("score"));
         CourseSectionRegistration newCourseSectionRegistration = courseSectionRegistrationRepository
-                .findByCourseSectionAndStudent(foundCourseSection.getId(), foundStudent.getStudentId())
-                .orElseThrow(CourseSectionRegistrationNotEmpty::new);
+                .findByCourseSectionAndStudent(courseSection.getId(), student.getStudentId())
+                .orElseThrow(CourseSectionRegistrationNotEmptyException::new);
         newCourseSectionRegistration.setScore(score);
         courseSectionRegistrationRepository.findById(newCourseSectionRegistration.getId()).map(courseSectionRegistration -> {
             courseSectionRegistration = newCourseSectionRegistration;
             return courseSectionRegistrationRepository.save(courseSectionRegistration);
         });
+        return "OK";
     }
+
+    private String foundError(String username, String password, CourseSection courseSection) {
+        if(!errorChecker.isUser(username, password)) throw new UnauthorisedException();
+        if(!ErrorChecker.isInstructor(username)) return "ERROR 403";
+        if(!ErrorChecker.isInstructorOfCourseSection(username, courseSection)) return "ERROR 403";
+        return null;
+    }
+
 }
