+++ src/main/java/tech/sobhan/golestan/services/UserService.java
@@ -4,15 +4,16 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import tech.sobhan.golestan.auth.User;
 import tech.sobhan.golestan.business.exceptions.UserNotFoundException;
+import tech.sobhan.golestan.business.exceptions.UnauthorisedException;
 import tech.sobhan.golestan.repositories.InstructorRepository;
 import tech.sobhan.golestan.repositories.StudentRepository;
 import tech.sobhan.golestan.repositories.UserRepository;
+import tech.sobhan.golestan.security.ErrorChecker;
 
 import java.util.List;
 
 import static tech.sobhan.golestan.security.config.PasswordConfiguration.getPasswordEncoder;
-import static tech.sobhan.golestan.utils.Util.createLog;
-import static tech.sobhan.golestan.utils.Util.deleteLog;
+import static tech.sobhan.golestan.utils.Util.*;
 
 @Service
 public class UserService {
@@ -20,39 +21,58 @@ public class UserService {
     private final UserRepository userRepository;
     private final StudentRepository studentRepository;//todo find usage
     private final InstructorRepository instructorRepository;
+    private final ErrorChecker errorChecker;
+
 
     public UserService(UserRepository userRepository,
-                       StudentRepository studentRepository, InstructorRepository instructorRepository) {
+                       StudentRepository studentRepository,
+                       InstructorRepository instructorRepository,
+                       ErrorChecker errorChecker) {
+        this.errorChecker = errorChecker;
         this.passwordEncoder = getPasswordEncoder();
         this.userRepository = userRepository;
         this.studentRepository = studentRepository;
         this.instructorRepository = instructorRepository;
     }
 
-    public List<User> list() {
-        //        if(allUsers.isEmpty()){
-//            throw new UserNotFoundException();
-//        }
+    public String list(String username, String password) {
+        errorChecker.checkIsUser(username, password);
+        return list().toString();
+    }
+
+    private List<User> list() {
         return userRepository.findAll();
     }
 
     public User create(String username, String password, String name, String phone, String nationalId) {
         User user = User.builder().username(username).password(passwordEncoder.encode(password)).name(name).phone(phone)
-                .nationalId(nationalId).active(false).admin(false).build();
+                .nationalId(nationalId).build();
         return create(user);
     }
 
     public User create(User user) {
         if (userExists(list(), user)) return null;
+        user.setActive(false);
+        user.setAdmin(false);
         createLog(User.class, user.getId());
         return userRepository.save(user);
     }
 
-    public User read(Long id) {
+    public String read(Long id, String username, String password) {
+        errorChecker.checkIsUser(username, password);
+        return read(id).toString();
+    }
+
+    private User read(Long id) {
         return userRepository.findById(id).orElseThrow(UserNotFoundException::new);
     }
 
-    public User update(User newUser, Long id) {
+    public String update(User newUser, Long id, String username, String password) {
+        errorChecker.checkIsAdmin(username, password);
+        return update(newUser, id).toString();
+    }
+
+    User update(User newUser, Long id) {
         userRepository.findById(id).map(user -> {
             user = newUser;
             return userRepository.save(user);
@@ -63,12 +83,14 @@ public class UserService {
         return newUser;
     }
 
-    public void delete(Long id) {
+    public String delete(Long id, String username, String password) {
+        errorChecker.checkIsAdmin(username, password);
         User user = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
         studentRepository.delete(user.getStudent());
         instructorRepository.delete(user.getInstructor());
         userRepository.delete(user);
         deleteLog(User.class, id);
+        return "OK";
     }
 
     private static boolean userExists(List<User> allUsers, User user) {
